new version now has all functionality the old one had
